refactor(dispatch): split cell_dispatch/interactions along message-family + entry-point seams
Behavior-preserving pure code movement — no logic or signature changes.

cell_dispatch/mod.rs (967): the single 40-arm CellToBaseMsg match is now a
thin two-level router. The outer match in mod.rs picks the message *family*;
each family's destructuring sub-match + arm bodies live in new sibling modules:
  - aoi_dispatch.rs          (AoI / space-lifecycle, deferred-buffer gate)
  - inventory_dispatch.rs    (inventory / bandolier / persisted-options)
  - vendor_dispatch.rs       (vendor store + two-party trade)
  - progression_dispatch.rs  (mission / grant / console / spawn / mail / minigame)
  - gate_teleport_dispatch.rs (gate-travel / reanchor / teleport)
A borrowed DispatchCtx bundles the shared transport/maps/pools. Existing
sibling helpers (aoi.rs, bandolier.rs, etc.) and the test modules are
untouched (tests drive through handle_cell_message).

cell/interactions/dispatch.rs (713): split into dispatch/ along the two entry
points — interact.rs (handle_interact) + initial_response.rs
(handle_initial_response) — with the inline #[cfg(test)] mod tests moved to
dispatch/tests.rs. MAX_INTERACT_DISTANCE stays pub(super) in dispatch/mod.rs
(consumed by the sibling loot module).

Behavior-preserving · re-export discipline · zero public-surface change

Part of #529
